--- example/out_forward_srv.conf
@@ -1,16 +1,0 @@
-<source>
-  @type dummy
-  tag test
-</source>
-
-<match test>
-  verify_connection_at_startup true
-  @type forward
-  expire_dns_cache 0
-  <server>
-
-    enable_dns_srv true
-    host bootjp.me
-    port 24224
-  </server>
-</match>
