+++ lib/fluent/plugin/out_forward.rb
@@ -974,6 +974,8 @@ module Fluent::Plugin
         end
       end
 
+      # @param host string
+      # @return [SRV]
       def resolve_srv(host)
         res = []
         adders = Resolv::DNS.new.getresources(host, Resolv::DNS::Resource::IN::SRV)
@@ -984,30 +986,87 @@ module Fluent::Plugin
         res
       end
 
+      # @param [SRV] srv_list
+      # @return [SRV]
       def srv_list_sort_priority_weight(srv_list)
-        if srv_list.nil? || srv_list.empty?
-          return []
+        return [] if srv_list.nil? || srv_list.empty?
+
+        # If the reply is NOERROR, ANCOUNT>0 and there is at least one
+        # SRV RR which specifies the requested Service and Protocol in
+        # the reply
+        sum = srv_list.inject(0) { |sum, srv| sum + srv.weight}
+        return srv_list if sum.zero?
+
+        # Sort the list by priority (lowest number first)
+        last = 0
+        current = 1
+
+        ret = []
+        srv_list.sort_by!(&:priority)
+        until (srv_list.size - 1) < current
+          if srv_list[last].priority != srv_list[current].priority
+            ret += weight_by_shuffle srv_list[last...current]
+            last = current
+          end
+          current += 1
+        end
+        ret += weight_by_shuffle srv_list[last...current]
+        ret
+      end
+
+      # @param [SRV] srv_list
+      # @return [SRV]
+      def weight_shuffle(srv_list)
+        return [] if srv_list.nil? || srv_list.empty?
+
+        # To select a target to be contacted next, arrange all SRV RRs
+        #  (that have not been ordered yet) in any order, except that all
+        # those with weight 0 are placed at the beginning of the list.
+        shuffle_start = nil
+        srv_list.sort_by!(&:weight)
+        srv_list.each_index do |index|
+          if srv_list[index].weight != 0
+            if shuffle_start.nil?
+              shuffle_start = index
+            end
+            if !shuffle_start.nil? && srv_list[index].weight != 0
+              shuffle_end = index
+              target = srv_list.dup
+              srv_list[shuffle_start..shuffle_end] = target[shuffle_start..shuffle_end].shuffle!
+            end
+          end
         end
 
-        srv_list.sort_by!(&:priority).chunk(&:priority).sort.each do |_, list|
-          sum = list.inject(0) { |sum, srv| sum + srv.weight}
-          while sum > 0 && list.count > 1 do
-            s = 0
-            select = Integer(rand(sum))
-            list.each_with_index do |srv, index|
-              s += srv.weight
-              if s > select
-                if index > 0
-                  list[0], list[index] = list[index], list[0]
-                end
-                break
-              end
-            end
-            sum -= list[0].weight
-          end
-          list
-        end
         srv_list
+      end
+
+      # @param [SRV] srv_list
+      # @return [SRV]
+      def weight_by_shuffle(srv_list)
+        return [] if srv_list.nil? || srv_list.empty?
+
+        # Compute the sum of the weights of those RRs, and with each RR
+        # associate the running sum in the selected order
+        sum = srv_list.inject(0) { |sum, srv| sum + srv.weight}
+
+        srv_list = weight_shuffle(srv_list)
+        ret = []
+        # Then choose a uniform random number between 0 and the sum computed (inclusive),
+        # and select the RR whose running sum value is the first in the selected order which is greater than or equal to the random number selected.
+        until sum <= 0 || srv_list.empty?
+          selector = Integer(rand(sum))
+          running_sum = 0
+
+          srv_list.each_index do |index|
+            running_sum += srv_list[index].weight
+            next unless running_sum > selector
+            ret.push srv_list[index]
+            sum -= srv_list[index].weight
+            srv_list.delete_at index
+            break
+          end
+        end
+        ret
       end
 
       def switch_original_host!
@@ -1030,19 +1089,19 @@ module Fluent::Plugin
           if @using_srv
             switch_original_host!
           end
-          @log.warn "srv record empty response ", host: host
+          @log.warn 'srv record empty response', host: host
           return
         end
 
         resp = srv_list_sort_priority_weight(resp)
-        available = resp.find {|s| s.available? }
+        available = resp.find {|s| s.available?}
 
         if available.nil?
           # unavailable is tried original host.
           if @using_srv
             switch_original_host!
           end
-          @log.warn "srv record does not available node. try using A record ", host: host
+          @log.warn 'srv record does not available node. try using A record', host: host
           return
         end
